+++ src/mons/aurox/IronWall.sol
@@ -25,7 +25,7 @@ contract IronWall is IMoveSet, BasicEffect {
         uint256 attackerPlayerIndex,
         uint256 attackerMonIndex,
         uint256,
-        uint240,
+        uint16,
         uint256
     ) external {
         // Check to see if the effect is already active
@@ -68,7 +68,7 @@ contract IronWall is IMoveSet, BasicEffect {
         return MoveClass.Self;
     }
 
-    function isValidTarget(IEngine, bytes32, uint240) external pure returns (bool) {
+    function isValidTarget(IEngine, bytes32, uint16) external pure returns (bool) {
         return true;
     }
