+++ test/mons/GhouliathTest.sol
@@ -150,7 +150,7 @@ contract GhouliathTest is Test, BattleHelper {
 
         // First move: Both players select their first mon (index 0)
         _commitRevealExecuteForAliceAndBob(
-            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint240(0), uint240(0)
+            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint16(0), uint16(0)
         );
 
         // Bob uses the attack (which KOs) on Alice's mon
@@ -170,7 +170,7 @@ contract GhouliathTest is Test, BattleHelper {
 
         // Alice swaps in mon index 1
         vm.startPrank(ALICE);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, 0, uint240(1), true);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, 0, uint16(1), true);
         engine.resetCallContext();
         // We wait for the REVIVAL_DELAY - 1 turns to pass
         for (uint256 i = 0; i < riseFromTheGrave.REVIVAL_DELAY() - 1; i++) {
@@ -188,12 +188,12 @@ contract GhouliathTest is Test, BattleHelper {
         assertEq(damageTaken, -99, "Alice's mon should have 1 HP");
 
         // Alice swaps in mon index 0, Bob does attack again, which KOs Alice's mon
-        _commitRevealExecuteForAliceAndBob(engine, commitManager, battleKey, SWITCH_MOVE_INDEX, 0, uint240(0), 0);
+        _commitRevealExecuteForAliceAndBob(engine, commitManager, battleKey, SWITCH_MOVE_INDEX, 0, uint16(0), 0);
 
         // Verify the mon is not revived after REVIVAL_DELAY turns
         // (First we swap in mon index 1)
         vm.startPrank(ALICE);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, 0, uint240(1), true);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, 0, uint16(1), true);
         engine.resetCallContext();
         for (uint256 i = 0; i < riseFromTheGrave.REVIVAL_DELAY() - 1; i++) {
             _commitRevealExecuteForAliceAndBob(
@@ -256,7 +256,7 @@ contract GhouliathTest is Test, BattleHelper {
 
         // First move: Both players select their first mon (index 0)
         _commitRevealExecuteForAliceAndBob(
-            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint240(0), uint240(0)
+            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint16(0), uint16(0)
         );
 
         // Bob uses the attack (which KOs) on Alice's mon
@@ -264,14 +264,14 @@ contract GhouliathTest is Test, BattleHelper {
 
         // Alice swaps in mon index 1
         vm.startPrank(ALICE);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, 0, uint240(1), true);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, 0, uint16(1), true);
         engine.resetCallContext();
         // Alice KOs Bob's mon
         _commitRevealExecuteForAliceAndBob(engine, commitManager, battleKey, 0, NO_OP_MOVE_INDEX, 0, 0);
 
         // Bob swaps in mon index 1
         vm.startPrank(BOB);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, 0, uint240(1), true);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, 0, uint16(1), true);
         engine.resetCallContext();
         // We wait for the REVIVAL_DELAY turns to pass
         for (uint256 i = 0; i < riseFromTheGrave.REVIVAL_DELAY() - 1; i++) {
@@ -346,7 +346,7 @@ contract GhouliathTest is Test, BattleHelper {
 
         // First move: Both players select their first mon (index 0)
         _commitRevealExecuteForAliceAndBob(
-            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint240(0), uint240(0)
+            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint16(0), uint16(0)
         );
 
         // Alice uses WitherAway on Bob's mon
@@ -448,7 +448,7 @@ contract GhouliathTest is Test, BattleHelper {
 
         // First move: Both players select their first mon (index 0)
         _commitRevealExecuteForAliceAndBob(
-            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint240(0), uint240(0)
+            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint16(0), uint16(0)
         );
 
         // Alice uses Osteoporosis on Bob's mon
@@ -500,12 +500,12 @@ contract GhouliathTest is Test, BattleHelper {
 
         // First move: Both players select their first mon (index 0)
         _commitRevealExecuteForAliceAndBob(
-            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint240(0), uint240(0)
+            engine, commitManager, battleKey, SWITCH_MOVE_INDEX, SWITCH_MOVE_INDEX, uint16(0), uint16(0)
         );
 
         // Alice does nothing, Bob switches to mon index 1
         _commitRevealExecuteForAliceAndBob(
-            engine, commitManager, battleKey, NO_OP_MOVE_INDEX, SWITCH_MOVE_INDEX, 0, uint240(1)
+            engine, commitManager, battleKey, NO_OP_MOVE_INDEX, SWITCH_MOVE_INDEX, 0, uint16(1)
         );
 
         // Alice uses Eternal Grudge on Bob's mon
